Log baton-do STDERR at debug rather than warn level
STDERR is out of band for extendo - all the error information is
transmitted via JSON on STDOUT. By default baton-do will print all
iRODS errors to STDERR (unless in --silent mode), including those we
are not interested in (e.g. file not found when probing for collections
to create). This makes STDERR warn logs somewhat alarmist.