Workaround for iRODS mkdir calls returning early
iRODS mkdir calls can return success to the client before the collection
has been created, leading to failures when e.g. the client then puts a
data object in the new collection.

This change introduces a short period of polling with a back-off, in an
attempt to confirm the presence of the collection before the
MakeCollection function returns. If the collection can't be detected
after 3 tries, an error is returned.