Improvements to sub-process shutdown.
Separated waitgroups for STDIN and STDOUT/STDERR of baton-do. Ensured
that STDIN is closed first and STDOUT/STDERR get a chance to read to
EOF before cmd.Wait() for the process to exit.

Launch each baton-do in a new process group to avoid being killed if
run in a shell and stopped with ^C.