Add iRODS 4.2.10
Tests to pass on iRODS 4.2.7 and 4.2.10.

Remove the conda-forge channel.

Use the convention of replResc as an iRODS test resource.

Add a BatonVersion function.

Add the JSON serialisation of Replicates in the Timestamp struct.

Add sorting of replicates in server responses.

Fix the JSON serialisation of Number in the Replicate struct.

Fix incorrect comparison of modified with created times in
SortTimestamps.

Fix some documentation strings.