Fix hang on baton client crash
Reorder the completion goroutine so that cmd.Wait() is called before waiting
for the IO goroutines.

Use client.IsRunning() to avoid data races checking for client state.

Add a regression test specifically for the cause of the original crash. This
will fail (now that we don't hang) and will pass once the baton client itself
if fixed.